requirements.txt
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
openpyxl==3.1.2
fastapi==0.109.2
uvicorn==0.27.0
streamlit==1.30.0
great-expectations==0.17.22
scikit-learn==1.3.2
python-dotenv==1.0.1